+++ examples/pwstrength.js
@@ -275,6 +275,7 @@ defaultOptions.ui.colorClasses = [
 defaultOptions.ui.showProgressBar = true;
 defaultOptions.ui.progressBarEmptyPercentage = 1;
 defaultOptions.ui.progressBarMinPercentage = 1;
+defaultOptions.ui.progressBarExtraCssClasses = '';
 defaultOptions.ui.showPopover = false;
 defaultOptions.ui.popoverPlacement = "bottom";
 defaultOptions.ui.showStatus = false;
@@ -380,20 +381,29 @@ var ui = {};
 
     ui.initProgressBar = function (options, $el) {
         var $container = ui.getContainer(options, $el),
-            progressbar = "<div class='progress'><div class='"; // Boostrap 2
+            progressbar = "<div class='progress ";
 
-        if (!options.ui.bootstrap2 && !options.ui.bootstrap4) {
+        if (options.ui.bootstrap2) {
+            // Boostrap 2
+            progressbar += options.ui.progressBarExtraCssClasses +
+                "'><div class='";
+        } else if (!options.ui.bootstrap2 && !options.ui.bootstrap4) {
             // Bootstrap 3
-            progressbar += "progress-";
+            progressbar += "'><div class='" +
+                options.ui.progressBarExtraCssClasses + " progress-";
         }
         progressbar += "bar'>";
+
         if (options.ui.bootstrap4) {
             // Boostrap 4
-            progressbar = "<progress class='progress' value='0' max='100'>";
+            progressbar = "<progress class='progress " +
+                options.ui.progressBarExtraCssClasses + "' value='0' max='100'>";
         }
+
         if (options.ui.showVerdictsInsideProgressBar) {
             progressbar += "<span class='password-verdict'></span>";
         }
+
         if (options.ui.bootstrap4) {
             progressbar += "</progress>";
         } else {